pyproject.toml
[project]
name = "billybjork-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.38",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "markdown>=3.9",
    "pillow>=11.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "uvicorn>=0.39.0",
]
